# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Solar Energy Management API",
        "description": "API documentation for Solar Energy Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List all projects",
                "description": "Get a list of all projects with optional filters",
                "operationId": "fbc21feeb3df8a2cb519efb6480ae7d9",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "Filter by client ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "uf",
                        "in": "query",
                        "description": "Filter by state (UF)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "client_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "uf": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "installation_type": {
                                                "type": "string",
                                                "example": "FIBROCIMENTO"
                                            },
                                            "equipment": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "type": {
                                                            "type": "string",
                                                            "example": "MODULO"
                                                        },
                                                        "quantity": {
                                                            "type": "integer",
                                                            "example": 10
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Create a new solar energy project",
                "operationId": "a9d5bd5914cf75eaebafcf8b9e320dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "uf",
                                    "installation_type",
                                    "equipment"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "SP"
                                    },
                                    "installation_type": {
                                        "description": "Valid values: FIBROCIMENTO, CERAMICO, METALICO, LAJE, SOLO",
                                        "type": "string",
                                        "example": "FIBROCIMENTO"
                                    },
                                    "equipment": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "type",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "type": {
                                                    "description": "Valid values: MODULO, INVERSOR, MICROINVERSOR, ESTRUTURA, STRINGBOX",
                                                    "type": "string",
                                                    "example": "MODULO"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "client_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "uf": {
                                            "type": "string"
                                        },
                                        "installation_type": {
                                            "type": "string"
                                        },
                                        "equipment": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}